# staticServer/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sense100 接口文档",
        "title": "sense100",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "740640431@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.5"
    },
    "host": "192.168.1.104:9494",
    "basePath": "/",
    "paths": {
...
        "/program/createOrUpdateMould": {
            "post": {
                "description": "创建模板",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "创建模板",
                "parameters": [
                    {
                        "description": "programMould",
                        "name": "programMould",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsProgramMould"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/program/createOrUpdateMouldCategory": {
            "post": {
                "description": "创建或更新模板类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "创建或更新模板类型",
                "parameters": [
                    {
                        "description": "mouldCategory",
                        "name": "mouldCategory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsMouldCategory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/program/delMould/{id}": {
            "delete": {
                "description": "删除模板信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "删除模板信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/program/delMouldCategory/{id}": {
            "delete": {
                "description": "根据Id删除类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "根据Id删除类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/program/getMouldCategoryOnParentId/{parentId}": {
            "get": {
                "description": "根据parentId获取类型子集",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "根据parentId获取类型子集",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "parentId",
                        "name": "parentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/program/getMouldInfo/{id}": {
            "get": {
                "description": "获取模板信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "获取模板信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsProgramMould"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/program/getMouldListOnCategoryId/{categoryId}": {
            "get": {
                "description": "根据类型获取模板集合",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "根据类型获取模板集合",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "categoryId",
                        "name": "categoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
...
        "/user/addBrowseRecord": {
            "post": {
                "description": "添加浏览记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "添加浏览记录",
                "parameters": [
                    {
                        "description": "addBrowseRecordReq",
                        "name": "addBrowseRecordReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.AddBrowseRecordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/user/doPraise": {
            "post": {
                "description": "添加点赞记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "添加点赞记录",
                "parameters": [
                    {
                        "description": "praise",
                        "name": "praise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsSpPraise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "登录 登录类型必传 手机号登录：手机号、密码必传  微信登录：微信id必传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "loginReq",
                        "name": "loginReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/request.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/user/registerUser": {
            "post": {
                "description": "注册用户 注册类型必传，手机号注册：手机号、密码必传 微信注册：微信id必传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "register",
                        "name": "registerUserReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/request.RegisterUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AddBrowseRecordReq": {
            "type": "object",
            "required": [
                "contentId",
                "type"
            ],
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "contentId": {
                    "description": "关联浏览内容Id，如：企业门面Id、小程序Id",
                    "type": "integer"
                },
                "createdTime": {
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "enabled": {
                    "type": "integer"
                },
                "headUrl": {
                    "description": "头像url",
                    "type": "string"
                },
                "iD": {
                    "description": "Id",
                    "type": "integer"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "nickName": {
                    "description": "昵称",
                    "type": "string"
                },
                "remarks": {
                    "description": "备注",
                    "type": "string"
                },
                "type": {
                    "description": "类型(1企业门面、2小程序)",
                    "type": "integer"
                },
                "wxId": {
                    "description": "微信Id",
                    "type": "string"
                }
            }
        },
...
        "model.SsMouldCategory": {
            "type": "object",
            "required": [
                "categoryName"
            ],
            "properties": {
                "categoryName": {
                    "description": "类型名称",
                    "type": "string"
                },
                "createdTime": {
                    "type": "string"
                },
                "creatorId": {
                    "description": "管理员Id",
                    "type": "integer"
                },
                "enabled": {
                    "type": "integer"
                },
                "iD": {
                    "type": "integer"
                },
                "parentId": {
                    "description": "关联父级Id",
                    "type": "integer"
                },
                "remarks": {
                    "type": "string"
                },
                "updatedTime": {
                    "type": "string"
                },
                "updatorId": {
                    "description": "管理员Id",
                    "type": "integer"
                }
            }
        },
        "model.SsProgramMould": {
            "type": "object",
            "required": [
                "categoryId",
                "name",
                "description",
                "salePrice",
                "status",
                "coverUrl",
                "detailPictureUrls"
            ],
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "coverUrl": {
                    "description": "封面url",
                    "type": "string"
                },
                "createdTime": {
                    "type": "string"
                },
                "creatorId": {
                    "description": "管理员Id",
                    "type": "integer"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "detailPictureUrls": {
                    "description": "详情图url,逗号分割",
                    "type": "string"
                },
                "enabled": {
                    "type": "integer"
                },
                "iD": {
                    "type": "integer"
                },
                "marketPrice": {
                    "description": "市场价",
                    "type": "number"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "salePrice": {
                    "description": "售价",
                    "type": "number"
                },
                "sourceUrl": {
                    "description": "源码url",
                    "type": "string"
                },
                "status": {
                    "description": "状态(1开发中、2已完成)",
                    "type": "integer"
                },
                "updatedTime": {
                    "type": "string"
                },
                "updatorId": {
                    "description": "管理员Id",
                    "type": "integer"
                }
            }
        },
...
        "model.SsSpPraise": {
            "type": "object",
            "properties": {
                "createdTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "enabled": {
                    "description": "0禁用、1启用",
                    "type": "integer"
                },
                "iD": {
                    "type": "integer"
                },
                "praiseContentId": {
                    "description": "关联点赞内容Id,如:关联点赞动态Id",
                    "type": "integer"
                },
                "type": {
                    "description": "类型(1点赞动态、2点赞资讯)",
                    "type": "integer"
                },
                "userId": {
                    "description": "关联用户Id",
                    "type": "integer"
                }
            }
        },
...
        "request.LoginReq": {
            "type": "object",
            "required": [
                "loginType"
            ],
            "properties": {
                "loginType": {
                    "description": "登录类型(1:手机号,2:微信)",
                    "type": "integer"
                },
                "password": {
                    "description": "密码(md5加密)",
                    "type": "string"
                },
                "userName": {
                    "description": "用户名",
                    "type": "string"
                },
                "wxId": {
                    "description": "微信id",
                    "type": "string"
                }
            }
        },
        "request.RegisterUserReq": {
            "type": "object",
            "properties": {
                "headUrl": {
                    "description": "头像url",
                    "type": "string"
                },
                "ipAddress": {
                    "description": "ip地址",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "nickName": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "type": {
                    "description": "1:手机号，2:微信",
                    "type": "integer"
                },
                "wxId": {
                    "description": "微信id",
                    "type": "string"
                }
            }
        },
...
        "response.BaseResponseSwag": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据jsonObject",
                    "type": "string"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                }
            }
        },
...
    }
}